requirements.txt
pytest
allure-pytest
selene --pre
faker
requests
